[Decomp][ObjectMgr] Cohesion split: extract locale string accessors to ObjectMgrLocaleStrings.cpp
